CSV-Implementation: Mcreated a BikeReportServiceImpl class which will handle the csv together with the CsvParserService. This method for now takes in the filepath of a the given CSV and will save it through the right DTO's.

+++ src/main/java/be/kdg/swiftby/service/dto/PedalDataDto.java
@@ -2,6 +2,6 @@ package be.kdg.swiftby.service.dto;
 
 public record PedalDataDto(
     double torqueCrank,
-    int cadence
+    double cadence
 ) {
 }
